Desktop: cable layout without crossings, remove unused mutex
- Sort(): pins of the right connectors are ordered so that the wires of
  one cable come to neighbour pins; pins of the left connectors follow
  the height of their wires on the right, so the wires do not cross
- Wires go straight through the cable blocks at the height of the right
  pin; cable blocks are placed around their wires without overlapping
- Old sort and block code removed from Cable (SortWiresRight,
  SortPinsLeft, CalcCableRect)
- Connector::GetPinPlace, MainCable::CollectWires shared with
  TesterPackage

+++ CableTesterDesktop/TesterWindow.hpp
@@ -105,7 +105,7 @@ public:
 					break;
 				}
 				if (!slot.used) break;              // the cables have no gaps
-				String name = String(slot.name, strnlen(slot.name, sizeof(slot.name)));
+				String name = String(slot.name, (int)strnlen(slot.name, sizeof(slot.name)));
 				names.Add(name);
 				slots.Add((i == info.currentSlot ? "► " : "  ") + IntStr(i), name,
 				          (int)slot.wireCount, (int)slot.netCount);
